% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.fit_BKMR_wasp.R
\name{kmbayes_Wasp}
\alias{kmbayes_Wasp}
\title{Fit scalable Bayesian kernel machine regression}
\usage{
kmbayes_Wasp(
  Z,
  X,
  y,
  n_subset,
  n_samp = 200,
  iter = 1000,
  varsel = FALSE,
  est.h = TRUE,
  Znew = NULL,
  knots = NULL,
  file_path = NULL,
  save_loc = FALSE,
  ...
)
}
\arguments{
\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{y}{a vector of outcome data of length \code{n}.}

\item{n_subset}{number of subsets to split the data}

\item{iter}{number of iterations to run the sampler}

\item{varsel}{TRUE or FALSE: indicator for whether to conduct variable selection on the Z variables in \code{h}}

\item{est.h}{TRUE or FALSE: indicator for whether to sample from the posterior distribution of the subject-specific effects h_i within the main sampler. This will slow down the model fitting.}

\item{Znew}{optional matrix of new predictor values at which to predict \code{h}, where each row represents a new observation. This will slow down the model fitting, and can be done as a post-processing step using \code{\link{SamplePred}}}

\item{knots}{optional matrix of knot locations for implementing the Gaussian predictive process of Banerjee et al (2008). Currently only implemented for models without a random intercept.}

\item{file_path}{working directory for saving the resulting rds files.}

\item{family}{a description of the error distribution and link function to be used in the model. Currently implemented for \code{gaussian} and \code{binomial} families.}

\item{id}{optional vector (of length \code{n}) of grouping factors for fitting a model with a random intercept. If NULL then no random intercept will be included.}

\item{verbose}{TRUE or FALSE: flag indicating whether to print intermediate diagnostic information during the model fitting.}

\item{starting.values}{list of starting values for each parameter. If not specified default values will be chosen.}

\item{control.params}{list of parameters specifying the prior distributions and tuning parameters for the MCMC algorithm. If not specified default values will be chosen.}

\item{groups}{optional vector (of length \code{M}) of group indictors for fitting hierarchical variable selection if varsel=TRUE. If varsel=TRUE without group specification, component-wise variable selections will be performed.}

\item{ztest}{optional vector indicating on which variables in Z to conduct variable selection (the remaining variables will be forced into the model).}

\item{rmethod}{for those predictors being forced into the \code{h} function, the method for sampling the \code{r[m]} values. Takes the value of 'varying' to allow separate \code{r[m]} for each predictor; 'equal' to force the same \code{r[m]} for each predictor; or 'fixed' to fix the \code{r[m]} to their starting values}
}
\value{
an object of class "bkmrfit", which has the associated methods:
\itemize{
  \item \code{\link{print}} (i.e., \code{\link{print.bkmrfit}})
  \item \code{\link{summary}} (i.e., \code{\link{summary.bkmrfit}})
}
}
\description{
Fits the Bayesian kernel machine regression (BKMR) model using Markov chain Monte Carlo (MCMC) methods with divide-and-conquer technique.
}
