% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPostmeanHnew_wasp.R
\name{postmeanhnew_wasp}
\alias{postmeanhnew_wasp}
\title{Compute the posterior mean and variance of \code{h} at a new predictor values}
\usage{
postmeanhnew_wasp(fit, X, y, Z, Znew = NULL, sel = NULL)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{y}{a vector of outcome data of length \code{n}.}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{Znew}{matrix of new predictor values at which to predict new \code{h}, where each row represents a new observation. If set to NULL then will default to using the observed exposures Z.}

\item{sel}{selects which iterations of the MCMC sampler to use for inference; see details}

\item{method}{method for obtaining posterior summaries at a vector of new points. now only support "approx", which is faster particularly for large datasets.}
}
\description{
Compute the posterior mean and variance of \code{h} at a new predictor values
}
